Model pt.2: add topic type "Translation" (#914).
"Translation" is a topic type.
It has Title, Language, and ISBN.
To create a Translation use the Create menu.

Associate a Work with a Translation by a generic Association. Create it manually.
Associate a Translation with its translator by (manually) creating a Work Involvement, and set its role to Translator.

The CROWD model should now be complete (in conjunction with the standard distro's dm4-contacts, dm4-events, and dm4-datetime models).

See #914.