Event page: show address w/ institution (#914).
If the Event's address is that of an Institution the address is displayed along with the institution name. The institution name is a link that leads to the institution page.

Usage hint: to reuse an institution's address as the event address reveal the institution's Address topic and associate it with the event. DM does the rest.

See #914